requirements.txt
async-generator==1.10
attrs==22.1.0
AutoChromedriver==2.2.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
Flask==2.2.2
h11==0.14.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
packaging==22.0
PyMySQL==1.0.2
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tiktok-scraper==1.0.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
webdriver-manager==3.8.5
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0
